requirements.txt
Flask
Flask-Cors
deep-translator
python-docx
PyPDF2
fpdf2
PyMuPDF
